Separate instance creation from model.
