+++ requirements.txt
@@ -1,4 +1,4 @@
-apscheduler
 flask
 flask-sqlalchemy
+Flask-APScheduler
 gunicorn
